Ref: create sentry flutter options and use generics
