+++ dart/lib/src/integration.dart
@@ -5,9 +5,9 @@ import 'sentry_options.dart';
 
 /// Code that provides middlewares, bindings or hooks into certain frameworks or environments,
 /// along with code that inserts those bindings and activates them.
-abstract class Integration {
+abstract class Integration<T extends SentryOptions> {
   /// A Callable method for the Integration interface
-  FutureOr<void> call(Hub hub, SentryOptions options);
+  FutureOr<void> call(Hub hub, T options);
 
   /// NoOp by default : only closeable integrations need to override
   void close() {}
