+++ dart/lib/src/sentry.dart
@@ -3,7 +3,6 @@ import 'dart:async';
 import 'default_integrations.dart';
 import 'hub.dart';
 import 'hub_adapter.dart';
-import 'integration.dart';
 import 'noop_isolate_error_integration.dart'
     if (dart.library.io) 'isolate_error_integration.dart';
 import 'noop_hub.dart';
@@ -13,7 +12,8 @@ import 'sentry_options.dart';
 import 'utils.dart';
 
 /// Configuration options callback
-typedef OptionsConfiguration = FutureOr<void> Function(SentryOptions);
+typedef OptionsConfiguration = FutureOr<void> Function<T extends SentryOptions>(
+    T);
 
 /// Runs a callback inside of the `runZonedGuarded` method, useful for running your `runApp(MyApp())`
 typedef AppRunner = FutureOr<void> Function();
@@ -31,30 +31,29 @@ class Sentry {
   /// passing a [AppRunner] callback allows to run the app within its own error zone (`runZonedGuarded`)
   /// https://api.dart.dev/stable/2.10.4/dart-async/runZonedGuarded.html
   static Future<void> init(
-    OptionsConfiguration optionsConfiguration, [
+    OptionsConfiguration optionsConfiguration, {
     AppRunner appRunner,
-    List<Integration> initialIntegrations,
-  ]) async {
+    SentryOptions options,
+  }) async {
     if (optionsConfiguration == null) {
       throw ArgumentError('OptionsConfiguration is required.');
     }
 
-    final options = SentryOptions();
-    await _initDefaultValues(options, appRunner, initialIntegrations);
+    final sentryOptions = options ?? SentryOptions();
+    await _initDefaultValues(sentryOptions, appRunner);
 
-    await optionsConfiguration(options);
+    await optionsConfiguration(sentryOptions);
 
-    if (options == null) {
+    if (sentryOptions == null) {
       throw ArgumentError('SentryOptions is required.');
     }
 
-    await _init(options);
+    await _init(sentryOptions);
   }
 
   static Future<void> _initDefaultValues(
     SentryOptions options,
     AppRunner appRunner,
-    List<Integration> initialIntegrations,
   ) async {
     // if no environment is set, we set 'production' by default, but if we know it's
     // a non-release build, or the SENTRY_ENVIRONMENT is set, we read from it.
@@ -70,19 +69,11 @@ class Sentry {
         ? const String.fromEnvironment('SENTRY_DSN')
         : options.dsn;
 
-    // we need to execute integrations in a specific order sometimes,
-    // and this initialIntegrations list makes it possible to inject and add
-    // integrations before the runZonedGuardedIntegration gets added
-    if (initialIntegrations != null && initialIntegrations.isNotEmpty) {
-      initialIntegrations
-          .forEach((integration) => options.addIntegration(integration));
-    }
-
     // Throws when running on the browser
     if (!isWeb) {
       // catch any errors that may occur within the entry function, main()
       // in the ‘root zone’ where all Dart programs start
-      options.addIntegration(IsolateErrorIntegration());
+      options.addIntegrationByIndex(0, IsolateErrorIntegration());
     }
 
     // finally the runZonedGuarded, catch any errors in Dart code running
