+++ flutter/lib/src/default_integrations.dart
@@ -3,11 +3,12 @@ import 'dart:async';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:sentry/sentry.dart';
+import 'sentry_flutter_options.dart';
 
 /// integration that capture errors on the FlutterError handler
-class FlutterErrorIntegration extends Integration {
+class FlutterErrorIntegration extends Integration<SentryFlutterOptions> {
   @override
-  void call(Hub hub, SentryOptions options) {
+  void call(Hub hub, SentryFlutterOptions options) {
     final defaultOnError = FlutterError.onError;
 
     FlutterError.onError = (FlutterErrorDetails errorDetails) async {
@@ -43,13 +44,13 @@ class FlutterErrorIntegration extends Integration {
 /// add an event processor to call a native channel method to load :
 /// - the device Contexts,
 /// - and the native sdk integrations and packages
-class LoadContextsIntegration extends Integration {
+class LoadContextsIntegration extends Integration<SentryFlutterOptions> {
   final MethodChannel _channel;
 
   LoadContextsIntegration(this._channel);
 
   @override
-  FutureOr<void> call(Hub hub, SentryOptions options) async {
+  FutureOr<void> call(Hub hub, SentryFlutterOptions options) async {
     options.addEventProcessor(
       (event, {hint}) async {
         try {
@@ -105,13 +106,13 @@ class LoadContextsIntegration extends Integration {
 }
 
 /// Enables Sentry's native SDKs (Android and iOS)
-class NativeSdkIntegration extends Integration {
+class NativeSdkIntegration extends Integration<SentryFlutterOptions> {
   final MethodChannel _channel;
 
   NativeSdkIntegration(this._channel);
 
   @override
-  FutureOr<void> call(Hub hub, SentryOptions options) async {
+  FutureOr<void> call(Hub hub, SentryFlutterOptions options) async {
     try {
       await _channel.invokeMethod<void>('initNativeSdk', <String, dynamic>{
         'dsn': options.dsn,
@@ -146,13 +147,14 @@ class NativeSdkIntegration extends Integration {
 }
 
 /// Loads the Android Image list for stack trace symbolication
-class LoadAndroidImageListIntegration extends Integration {
+class LoadAndroidImageListIntegration
+    extends Integration<SentryFlutterOptions> {
   final MethodChannel _channel;
 
   LoadAndroidImageListIntegration(this._channel);
 
   @override
-  FutureOr<void> call(Hub hub, SentryOptions options) {
+  FutureOr<void> call(Hub hub, SentryFlutterOptions options) {
     options.addEventProcessor(
       (event, {hint}) async {
         try {
